speedtest: new package to measure network performance
Package the betterspeedtest.sh and netperfrunner.sh scripts for easier use
on OpenWrt platforms. The first script emulates a web-based speed test by
downloading and then uploading from an internet server. The second script
performs a simultaneous download and upload, and simulates the FLENT test
program.

These scripts were spawned by the CeroWrt project and used to measure
network throughput and latency, as part of overall bufferbloat mitigation.
They are included here with the permission of the author, Rich Brown.

Signed-off-by: Tony Ambardar <itugrok@yahoo.com>