Update README.md
I made another editorial pass over the document. 

I changed a lot of text, both to make the most important points stand out, and to move a lot of details farther down. I also changed the "viewpoint" to support a newcomer to Travelmate.

I believe that everything remains correct, but you'll need to read it carefully to make sure that I am not telling falshoods...

And I still don't know how to commit to master, so I'll ask you to proofread this , then check it in.

I hope you're enjoying your holiday!